feat: nuke-trail cosmetic effect + tabbed effects picker
Add a "nukeTrail" effectType alongside transportShipTrail — coloring nuke
trails independently from boat trails — and organize the effects picker and
store into per-effectType tabs.

Rendering: boat and nuke trails share one trail texture, so an independent
color needs a per-tile unit-class signal. The trail texture is promoted
R8UI -> R16UI (owner in bits 0-11, nuke bit at bit 12; also lifting the old
255-owner truncation), and the per-player effect texture is widened to two
stacked blocks (block 0 = transportShipTrail, block 1 = nukeTrail). The shader
masks the owner, derives the block from the nuke bit, and reuses the existing
gradient/transition decode.

Schema: the shared attributes schema is renamed TrailEffectAttributesSchema;
NukeTrailEffectSchema joins EffectSchema and CosmeticsSchema.effects.
Everything downstream (server validation, selection, picker) iterates
EFFECT_TYPES, so it handles the new type with no per-type code.

UI: the selection modal shows one tab per effectType (BaseModal tabs); the
store's EFFECTS panel gets an internal sub-tab bar (its top-level tabs can't
nest). Tabs are always present, so a type you own entirely still shows as an
empty tab.

+++ src/client/render/frame/Upload.ts
@@ -17,10 +17,13 @@ import type {
  * Satisfied by GameView through TypeScript structural typing.
  */
 export interface FrameUploadTarget {
-  uploadTileAndTrailState(tileState: Uint16Array, trailState: Uint8Array): void;
+  uploadTileAndTrailState(
+    tileState: Uint16Array,
+    trailState: Uint16Array,
+  ): void;
   uploadLiveDelta(tileState: Uint16Array, changedTiles: TilePair[]): void;
   uploadLiveTrailDelta(
-    trailState: Uint8Array,
+    trailState: Uint16Array,
     dirtyRowMin: number,
     dirtyRowMax: number,
   ): void;
